refactor: extract PDF table scope grid geometry
